case_study_sources/requirements.txt
attrs==21.2.0
certifi==2020.4.5.1
chardet==3.0.4
charset-normalizer==2.0.12
configparser==5.0.0
coverage==5.5
defusedxml==0.6.0
dj-rest-auth==2.1.3
Django==2.2
django-allauth==0.42.0
django-bootstrap-modal-forms==1.5.0
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-crispy-forms==1.9.1
django-debug-toolbar==2.2
django-js-asset==1.2.2
django-recaptcha==2.0.6
django-sendgrid-v5==0.9.1
django-taggit==1.3.0
django-tempus-dominus==5.1.2.16
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
future==0.18.2
httpie==2.6.0
idna==2.9
importlib-metadata==4.6.1
iniconfig==1.1.1
oauthlib==3.1.0
packaging==21.0
Pillow==7.1.2
pluggy==0.13.1
psycopg2==2.8.5
py==1.10.0
Pygments==2.12.0
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.4
python-dateutil==2.8.2
python-decouple==3.3
python-http-client==3.3.2
python-magic-bin==0.4.14
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
sendgrid==6.6.0
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
starkbank-ecdsa==1.1.0
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.25.9
zipp==3.5.0
